% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI_function.R
\name{BMI}
\alias{BMI}
\title{Calcualte BMI}
\usage{
BMI(height, weight)
}
\arguments{
\item{height}{numeric nubmer}

\item{weight}{nubmerc number}
}
\value{
Numeric number
}
\description{
The BMI function is a simple function to assess an individual's health. It incorporates two variables, height(meters) and weight (kilograms).
}
\examples{

BMI(1.70,50)


}
