% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon_cancer_female_function.R
\name{colon_cancer_female}
\alias{colon_cancer_female}
\title{Colon cancer prediction for females}
\usage{
colon_cancer_female(age, race, education, estrogen, diabetes, smoking,
  family_history, multivitamins, BMI, NSAIDS, drinks)
}
\arguments{
\item{age}{numberic number}

\item{race}{numeric number}

\item{education}{numeric number}

\item{estrogen}{numeric number}

\item{diabetes}{numeric number}

\item{smoking}{numeric number}

\item{family_history}{numeric number}

\item{multivitamins}{numeric number}

\item{BMI}{numeric number}

\item{NSAIDS}{numeric number}

\item{drinks}{numeric number}
}
\value{
numberic number
}
\description{
The colon cancer prediction function is a function used to calculate colon cancer risk for females. It is based on a published study by Wells et al., 2014. The function uses parameters that have been correlated to colon cancer such as age, physical activity,among many other ourlined below.
}
\examples{
colon_cancer_female(50,3,16,2,1,0,0,0,33.5,0,0)
}
