% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon_cancer_male_function.R
\name{colon_cancer_male}
\alias{colon_cancer_male}
\title{Colon cancer prediction for males}
\usage{
colon_cancer_male(age, race, education, aspirin, diabetes, smoking,
  family_history, multivitamins, BMI, red_meat, drinks,
  physical_activiites)
}
\arguments{
\item{age}{numberic number}

\item{race}{numeric number}

\item{education}{numeric number}

\item{aspirin}{numeric number}

\item{diabetes}{numeric number}

\item{smoking}{numeric number}

\item{family_history}{numeric number}

\item{multivitamins}{numeric number}

\item{BMI}{numeric number}

\item{red_meat}{numeric number}

\item{drinks}{numeric number}

\item{physical_activiites}{nuemric number}
}
\value{
numberic number
}
\description{
The colon cancer prediction function is a function used to calculate colon cancer risk for males. It is based on a published study by Wells et al., 2014. The function uses parameters that have been correlated to colon cancer such as age, physical activity,among many other ourlined below. amomg
}
\examples{
colon_cancer_male(60,1,12,1,0,20,0,1,20,1,2,0)

}
