% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungcancer_function.R
\name{lungcancer}
\alias{lungcancer}
\title{Predict absolute risk of lung cancer}
\usage{
lungcancer(age, sex, smoking, pneumonia, asbestos, malignant_tumour,
  family_history)
}
\arguments{
\item{age}{Numeric number}

\item{sex}{Numeric number}

\item{smoking}{Numeric number}

\item{pneumonia}{Numeric number}

\item{asbestos}{Numeric number}

\item{malignant_tumour}{Numeric number}

\item{family_history}{Numeric number}
}
\value{
Numeric number
}
\description{
This function calculates the risk of developing lung cancer within 5-years. It is based on the LLP model of lung cancer and uses numerous parameters and risk  factors that been show to impact lung cancer. These variables include sex, tobacco use, exposures, among many other varianles, shown below.
}
\examples{

lungcancer(65,2,37,1,0,0,70)

}
